Drop migrations and start over with IDs type change

+++ Website/Migrations/ApplicationDbContextModelSnapshot.cs
@@ -226,14 +226,14 @@ namespace Website.Migrations
 
             modelBuilder.Entity("Website.Models.Game", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
-                    b.Property<int>("GenreId")
-                        .HasColumnType("int");
+                    b.Property<long>("GenreId")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("ImageUrl")
                         .HasMaxLength(255)
@@ -260,11 +260,11 @@ namespace Website.Migrations
 
             modelBuilder.Entity("Website.Models.Genre", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Name")
                         .IsRequired()
@@ -283,19 +283,19 @@ namespace Website.Migrations
 
             modelBuilder.Entity("Website.Models.Review", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"));
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Body")
                         .IsRequired()
                         .HasMaxLength(2147483647)
                         .HasColumnType("text");
 
-                    b.Property<int>("GameId")
-                        .HasColumnType("int");
+                    b.Property<long>("GameId")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("Title")
                         .IsRequired()
